Conectar a BD 65 - IniciodeSesion4 - ErrorConcurrencia

+++ PrototipoOT/frmInicioSesion.cs
@@ -20,7 +20,7 @@ namespace PrototipoOT
 
         private void button1_Click(object sender, EventArgs e)
         {
-            if (this.cUENTAS_DE_USUARIOTableAdapter.ExistsAccount(txtNombre.Text, txtContraseña.Text) == 1)
+            if ((int) this.cUENTAS_DE_USUARIOTableAdapter.ExistsAccount(txtNombre.Text, txtContraseña.Text) == 1)
             {
                 int permiso = (int) this.cUENTAS_DE_USUARIOTableAdapter.FetchPermiso(txtNombre.Text);
